feat(Compiler): separate compiler components
