+++ packages/hedgehog-lab/src/components/Compiler/RCompilerStates.ts
@@ -1,0 +1,19 @@
+import {atom} from "recoil";
+
+export const compilerLoadingState = atom({
+    key: 'compilerLoading',
+    default: false
+})
+
+export const compilerResultState = atom({
+    key: 'compilerResult',
+    default: {
+        outputItem: [],
+        outputString: ''
+    }
+})
+
+export const compilerReFetchState = atom({
+    key: 'compilerReFetch',
+    default: false
+})
