+++ packages/hedgehog-lab/src/pages/Main/Main.tsx
@@ -7,9 +7,10 @@ import Footer from "../../components/DeprecatedCode/Footer/Footer";
 import YourCode from "../../components/YourCode/YourCode";
 import {queryCache} from "react-query";
 import {useSetRecoilState} from "recoil";
-import {compilerReFetchState, editorCodeState} from "../../components/YourCode/RYourCodeStates";
+import {editorCodeState} from "../../components/YourCode/RYourCodeStates";
 import {useParams} from "react-router-dom";
 import {tutorials} from "../../tutorials";
+import {compilerReFetchState, compilerResultState} from "../../components/Compiler/RCompilerStates";
 
 const DEFAULT_SOURCE = `//write your code here
 print("hello world")
@@ -21,6 +22,7 @@ const lastRunningCode = localStorage.getItem('lastRunningCode')
 const Main = (): React.ReactElement => {
     const setEditorCode = useSetRecoilState<string>(editorCodeState)
     const setCompilerReFetch = useSetRecoilState<boolean>(compilerReFetchState);
+    const setCompilerResult = useSetRecoilState(compilerResultState)
     // If auto_run=true, then hedgehog lab will run the script automatically after loading the code
     const [autoRun, setAutoRun] = useState<boolean>(false)
 
@@ -57,9 +59,15 @@ const Main = (): React.ReactElement => {
     // Temporary, set up route for tutorial
     const {tutorialID} = useParams()
     useEffect(() => {
-        const tutorialDetail = tutorials.find(o => o.description === tutorialID)
-        setEditorCode(tutorialDetail ? tutorialDetail.source : '')
-        setCompilerReFetch(true)
+        if (tutorialID) {
+            setCompilerResult({
+                outputItem: [],
+                outputString: ''
+            });
+
+            const tutorialDetail = tutorials.find(o => o.description === tutorialID)
+            setEditorCode(tutorialDetail ? tutorialDetail.source : '')
+        }
     }, [tutorialID])
 
     return (
